actor/v7pushaction/transform_manifest_with_start_command_flag.go
package v7pushaction

import (
	"code.cloudfoundry.org/cli/command/translatableerror"
	"code.cloudfoundry.org/cli/util/manifestparser"
)

func TransformManifestWithStartCommandFlag(manifest manifestparser.ParsedManifest, overrides FlagOverrides) (manifestparser.ParsedManifest, error) {
	if overrides.StartCommand.IsSet {
		if manifest.ContainsMultipleApps() {
			return manifest, translatableerror.CommandLineArgsWithMultipleAppsError{}
		}

		webProcess := manifest.GetFirstAppWebProcess()
		if webProcess != nil {
			webProcess.StartCommand = overrides.StartCommand.Value
		} else {
			app := manifest.GetFirstApp()
			app.StartCommand = overrides.StartCommand.Value
		}
	}

	return manifest, nil
}
